+++ app/templates/_gulpfile.js
@@ -3,7 +3,7 @@
 var gulp = require('gulp');
 var gulpUtil = require('gulp-util');
 var del = require('del');
-var webpack = require('gulp-webpack');
+var webpack = require('webpack-stream');
 var named = require('vinyl-named');
 var minifyHtml = require('gulp-minify-html');
 var uglify = require('gulp-uglify');
@@ -15,8 +15,10 @@ var modRewrite = require('connect-modrewrite');
 var runSequence = require('run-sequence');
 var karma = require('gulp-karma');
 var opn = require('opn');
-var webpackStatsHelper = require('./webpack.stats.helper.js');
+var webpackStatsHelper = require('./webpack.stats.helper');
 var replace = require('gulp-replace-task');
+var proxy = require('proxy-middleware');
+var url = require('url');
 
 var isOpen = false;
 var autoprefixerBrowsers = [
@@ -45,6 +47,12 @@ var webpackStatsOptions = {
   source: false,
   errorDetails: false
 };
+var proxyOptions = [
+  //{
+  //  endpoint: 'http://localhost:8000/api',
+  //  route: '/api'
+  //}
+];
 
 function openApp() {
   if (!isOpen) {
@@ -61,6 +69,13 @@ function handleWebpack(error, stats) {
     gulpUtil.log(stats.toString(webpackStatsOptions));
     openApp();
   }
+}
+
+function createProxyOption(endpoint, route) {
+  var proxyOption = url.parse(endpoint);
+  proxyOption.route = route;
+  proxyOption.rejectUnauthorized = false;
+  return proxyOption;
 }
 
 gulp.task('clean', function () {
@@ -137,15 +152,17 @@ gulp.task('copy', function () {
 });
 
 gulp.task('browserSync', function (callback) {
+  var middleware = proxyOptions.map(function (proxyOption) {
+    return proxy(createProxyOption(proxyOption.endpoint, proxyOption.route));
+  });
+  middleware.push(modRewrite([
+    '!\\.\\w+$ /index.html [L]'
+  ]));
   browserSync({
-    files: ['app/*.html', '.tmp/**/*'],
+    files: ['app/*.html', '.tmp/**/*', '!.tmp/**/*.css'],
     server: {
       baseDir: ['.tmp', 'app'],
-      middleware: [
-        modRewrite([
-          '!\\.\\w+$ /index.html [L]'
-        ])
-      ]
+      middleware: middleware
     },
     port: 3000,
     reloadOnRestart: false,
@@ -171,14 +188,16 @@ gulp.task('build', function (callback) {
 });
 
 gulp.task('serve:dist', ['build'], function () {
+  var middleware = proxyOptions.map(function (proxyOption) {
+    return proxy(createProxyOption(proxyOption.endpoint, proxyOption.route));
+  });
+  middleware.push(modRewrite([
+    '!\\.\\w+$ /index.html [L]'
+  ]));
   browserSync({
     server: {
       baseDir: 'dist',
-      middleware: [
-        modRewrite([
-          '!\\.\\w+$ /index.html [L]'
-        ])
-      ]
+      middleware: middleware
     }
   });
 });